+++ termux_tts/__init__.py
@@ -29,7 +29,7 @@ from .exceptions import (
 ONNXNeuralEngine = SherpaNeuralEngine
 ONNXResult = SherpaResult
 
-__version__ = "1.4.2"
+__version__ = "1.4.3"
 __all__ = [
     "TTSEngine",
     "load",